Перенес создание компонента фильтрации и сортировки в page-controller.js
